refactor(minimum_rule_based_planner): merge obstacle_slow_down_plan back into obstacle_slow_down_utils
The plan stage was split out into its own file while it was being reworked.
Now that it has settled, move it back so the definitions sit next to the ones
they share the header with, ordered as the header declares them.
